requirement.txt
iniconfig==2.0.0
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
pytesseract==0.3.13
pytest==8.3.3
PyYAML==6.0.2
